Extract funding-finder and programmes into standalone routers
